moving all utilities out of its own folder
