+++ docs_src/routes/routes.ts
@@ -12,9 +12,9 @@ import PillNavigation from '../components/PillNavigation.svelte';
 import Spinner from '../components/Spinner.svelte';
 import Toggler from '../components/Toggler.svelte';
 import Tooltip from '../components/Tooltip.svelte';
-import Animation from '../utilities/Animation/Animation.svelte';
-import DataTheme from '../utilities/DataTheme/DataTheme.svelte';
-import Flex from '../utilities/Flex/Flex.svelte';
+import Animation from '../utilities/Animation.svelte';
+import DataTheme from '../utilities/DataTheme.svelte';
+import Flex from '../utilities/Flex.svelte';
 import Fonts from '../utilities/Fonts.svelte';
 import Grid from '../utilities/Grid.svelte';
 import Helpers from '../utilities/Helpers.svelte';
