+++ docs_src/utilities/Animation.svelte
@@ -1,7 +1,7 @@
 <script>
   import Prism from 'svelte-prism';
 
-  import { ArticleCard, Badge, PillNavigation, Pill, PillContent, PillList } from '../../../src/';
+  import { ArticleCard, Badge, PillNavigation, Pill, PillContent, PillList } from '../../src/';
 
   let article = {
     href: '#',
